modified:   DnsSwift/src/DnsSwift2Kifer/DnsKicker.java 	modified:   DnsSwift/src/DnsSwift2Kifer/DnsOperator.java 	modified:   DnsSwift/src/DnsSwift2Kifer/DnsRun.java 	modified:   DnsSwift/src/DnsSwift2Kifer/DnsRunFull.java 	modified:   DnsSwift/src/DnsSwift2Kifer/DnsRunStatistics.java 	new file:   DnsSwift/src/DnsSwift2Kifer/HookShutdown.java 	modified:   DnsSwift/src/DnsSwift2Kifer/LookupBuilder.java 	new file:   DnsSwift/src/DnsSwift2Kifer/TalkToSystem.java 	modified:   DnsSwift/src/DnsSwift2Kifer/TryCode1.java 	new file:   out/production/DnsSwift/DnsSwift2Kifer/DnsKicker$1.class 	modified:   out/production/DnsSwift/DnsSwift2Kifer/DnsKicker.class 	modified:   out/production/DnsSwift/DnsSwift2Kifer/DnsOperator.class 	modified:   out/production/DnsSwift/DnsSwift2Kifer/DnsRun.class 	modified:   out/production/DnsSwift/DnsSwift2Kifer/DnsRunFull.class 	modified:   out/production/DnsSwift/DnsSwift2Kifer/DnsRunStatistics.class 	modified:   out/production/DnsSwift/DnsSwift2Kifer/LookupBuilder.class 	new file:   out/production/DnsSwift/DnsSwift2Kifer/TalkToSystem.class 	modified:   out/production/DnsSwift/DnsSwift2Kifer/TryCode1.class 	deleted:    .idea/DnsSwift2Kifer.iml 	modified:   .idea/compiler.xml 	deleted:    .idea/inspectionProfiles/Project_Default.xml 	deleted:    .idea/inspectionProfiles/profiles_settings.xml 	modified:   .idea/misc.xml 	modified:   .idea/modules.xml 	deleted:    .idea/vcs.xml 	modified:   .idea/workspace.xml 	.idea/artifacts/

+++ DnsSwift/src/DnsSwift2Kifer/DnsRun.java
@@ -23,7 +23,7 @@ public abstract class DnsRun implements Runnable {
 
     public void setTotalRequests(int totalRequests){this.totalRequests = totalRequests;}
 
-    public void setSleepTime(int sleepTime){this.sleepTime = sleepTime;}
+    public void setSleepTime(long sleepTime){this.sleepTime = sleepTime;}
 
     public void setDebugMode(boolean debugMode){this.debugMode=debugMode;}
